added pagination

+++ repositories/contacts.js
@@ -1,25 +1,44 @@
 const Contact = require("../model/contact");
 
-const listContacts = async () => {
-  const results = await Contact.find();
+const listContacts = async (userId, query) => {
+  // const results = await Contact.find({ owner: userId }).populate({
+  //   path: "owner",
+  //   select: "email subscription",
+  // });
+  const {
+    sortBy,
+    sortByDesc,
+    filter,
+    favorite = false,
+    limit = 5,
+    offset,
+  } = query;
+  const optionsSearch = { owner: userId };
+  if (favorite) {
+    optionSearch.favorite = favorite;
+  }
+  const results = Contact.paginate(optionsSearch, { limit, offset });
   return results;
 };
 
-const getContactById = async (contactId) => {
-  return await Contact.findOne({ _id: contactId });
+const getContactById = async (userId, contactId) => {
+  return await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: "owner",
+    select: "email subscription",
+  });
 };
 
-const removeContact = async (contactId) => {
-  return await Contact.findOneAndRemove({ _id: contactId });
+const removeContact = async (userId, contactId) => {
+  return await Contact.findOneAndRemove({ _id: contactId, owner: userId });
 };
 
-const addContact = async (body) => {
-  return await Contact.create(body);
+const addContact = async (userId, body) => {
+  return await Contact.create({ ...body, owner: userId });
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true }
   );
